+++ lib/mongory/matchers/and_matcher.rb
@@ -7,7 +7,7 @@ module Mongory
     # It evaluates an array of subconditions and returns true only if *all* of them match.
     #
     # Unlike other matchers, AndMatcher flattens the underlying matcher tree by
-    # delegating each subcondition to a `ConditionMatcher`, and further extracting
+    # delegating each subcondition to a `HashConditionMatcher`, and further extracting
     # all inner matchers. Duplicate matchers are deduplicated by `uniq_key`.
     #
     # This allows the matcher trace (`.explain`) to render as a flat list of independent conditions.
@@ -21,20 +21,20 @@ module Mongory
     #
     # @see AbstractMultiMatcher
     class AndMatcher < AbstractMultiMatcher
-      # Constructs a ConditionMatcher for each subcondition.
+      # Constructs a HashConditionMatcher for each subcondition.
       # Conversion is disabled to avoid double-processing.
       enable_unwrap!
 
       # Returns the flattened list of all matchers from each subcondition.
       #
-      # Each condition is passed to a ConditionMatcher, then recursively flattened.
+      # Each condition is passed to a HashConditionMatcher, then recursively flattened.
       # All matchers are then deduplicated using `uniq_key`.
       #
       # @return [Array<AbstractMatcher>]
       # @see AbstractMatcher#uniq_key
       define_instance_cache_method(:matchers) do
         @condition.flat_map do |condition|
-          ConditionMatcher.new(condition).matchers
+          HashConditionMatcher.new(condition).matchers
         end.uniq(&:uniq_key)
       end