+++ lib/mongory/query_builder.rb
@@ -91,36 +91,12 @@ module Mongory
       self.and('$or' => conditions)
     end
 
-    # @deprecated Will be removed in v2.0.0. Sort externally instead.
-    # Sorts the records by the given keys in ascending order.
-    #
-    # @param sort_keys [Array<Symbol, String>]
-    # @return [QueryBuilder]
-    def asc(*sort_keys)
-      warn '[Mongory] `Mongory::QueryBuilder#asc` is deprecated and will be removed in v2.0.0.' \
-           'Please sort outside Mongory.'
-
-      dup_instance_exec do
-        @records = sort_by_keys(sort_keys) do |a, b|
-          a <=> b
-        end
-      end
+    def in(condition)
+      self.and(wrap_values_with_key(condition, '$in'))
     end
 
-    # @deprecated Will be removed in v2.0.0. Sort externally instead.
-    # Sorts the records by the given keys in descending order.
-    #
-    # @param sort_keys [Array<Symbol, String>]
-    # @return [QueryBuilder]
-    def desc(*sort_keys)
-      warn '[Mongory] `Mongory::QueryBuilder#desc` is deprecated and will be removed in v2.0.0.' \
-           'Please sort outside Mongory.'
-
-      dup_instance_exec do
-        @records = sort_by_keys(sort_keys) do |a, b|
-          b <=> a
-        end
-      end
+    def nin(condition)
+      self.and(wrap_values_with_key(condition, '$nin'))
     end
 
     # Limits the number of records returned by the query.
@@ -157,15 +133,14 @@ module Mongory
     alias_method :selector, :condition
 
     # Prints the internal matcher tree structure for the current query.
-    # Uses `PP` to output a human-readable visual tree of matchers.
+    # Will output a human-readable visual tree of matchers.
     # This is useful for debugging and visualizing complex conditions.
     #
     # @return [void]
     def explain
-      @matcher.match(@records.first)
-      pp = PP.new($stdout)
-      @matcher.render_tree(pp)
-      pp.flush
+      @matcher.match?(@records.first)
+      @matcher.render_tree
+      nil
     end
 
     private
@@ -203,16 +178,9 @@ module Mongory
       set_matcher(condition_dup)
     end
 
-    # @private
-    # Sorts a list of records by one or more keys.
-    #
-    # @param sort_keys [Array]
-    # @yield [a, b] comparison block
-    # @return [Array]
-    def sort_by_keys(sort_keys)
-      sort do |a, b|
-        yield sort_keys.map { |key| a[key] },
-              sort_keys.map { |key| b[key] }
+    def wrap_values_with_key(condition, key)
+      condition.transform_values do |sub_condition|
+        { key => sub_condition }
       end
     end
   end